backend/requirements.txt
fastapi[standard]>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
bcrypt==3.2.2
